requirements.txt
hydra-core
omegaconf
numpy
matplotlib
pytest